requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Google Gemini
google-generativeai>=0.3.0

# Vector DB and Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
pypdf>=3.17.0
python-docx>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
datasets>=2.14.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0

